Align maximum migration name length type in exception with attribute

+++ src/Puro/Exceptions/MigrationNameTooLongException.cs
@@ -2,7 +2,7 @@
 
 public class MigrationNameTooLongException : PuroException
 {
-	public MigrationNameTooLongException(short maximumLength)
+	public MigrationNameTooLongException(int maximumLength)
 		: base($"Maximum migration name length is {maximumLength} characters.")
 	{
 	}
